create oscillator and gain in AudioContextManager, add parameters for play in ChilloutAudio
