+++ src/index.ts
@@ -2,6 +2,7 @@ import TracksManager from './modules/TracksManager';
 import Track from './modules/Track';
 import SineWaveInstrument from './modules/SineWaveInstrument';
 import { Melody } from './modules/Melody';
+import { Instruments } from './types/instruments';
 
 /**
  * Chillout audio class
@@ -30,11 +31,17 @@ export default class ChilloutAudio {
 
   /**
    * Method for start playing melody
+   * @param notes {String} - notes in melody
+   * @param instrument {Instruments} - name of instrument
    */
-  public play(): void {
-    const melody = new Melody('A4 A5 D3 e4 d4 f8 a6 b3 c2 A5 A4');
+  public play(notes: string = 'A4 A5 D3 E4', instrument: Instruments = Instruments.SINE_WAVE_INSTRUMENT): void {
+    const melody = new Melody(notes);
 
-    this.track = new Track(new SineWaveInstrument(), melody);
+    switch (instrument) {
+      case Instruments.SINE_WAVE_INSTRUMENT:
+        this.track = new Track(new SineWaveInstrument(), melody);
+        break;
+    }
     this.track.play();
   }
